# docs/openapi/core-api.yaml
openapi: 3.0.3
info:
  title: CryptoPay Core API
  version: 1.0.0
servers:
  - url: http://localhost:3001
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: idempotency-key
      in: header
      required: true
      schema:
        type: string
...
paths:
...
  /internal/v1/ops/reconciliation/runs/{runId}:
    get:
      security:
        - bearerAuth: []
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reconciliation run details
  /internal/v1/ops/reconciliation/issues:
    get:
      security:
        - bearerAuth: []
      parameters:
        - name: since
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Reconciliation issues list
  /internal/v1/ops/reconciliation/run:
    post:
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                outputPath:
                  type: string
      responses:
        '202':
          description: Reconciliation run queued
  /internal/v1/ops/jobs/retention/run:
    post:
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Retention job completed
  /internal/v1/ops/jobs/key-verification/run:
    post:
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Key verification completed
...